requirements.txt
livekit-agents[google,openai,elevenlabs,silero,turn-detector]==1.2.*
livekit-plugins-noise-cancellation==0.2.*
livekit-plugins-langchain==1.1.*
python-dotenv
langchain-openai
langgraph
pydantic